Implemented PointCloudVisualizer with RGB stream
